feat: Implement genre filtering and movie rating

+++ MovieTheater.Domain/Entities/Movie.cs
@@ -25,7 +25,7 @@ namespace MovieTheater.Domain.Entities
         public string PosterImage { get; set; }
         public string BackgroundImage { get; set; }
         public MovieStatus Status { get; set; }       // Trạng thái phim
-        public float Rating { get; set; }             // Điểm đánh giá
+        public double Rating { get; set; }             // Điểm đánh giá
 
         // Navigation
         public ICollection<ShowTime> Showtimes { get; set; }
